requirements.txt
fastapi==0.68.1
uvicorn[standard]==0.15.0
GitPython==3.1.24
aiofiles==0.7.0
Authlib==0.15.4
httpx==0.19.0
itsdangerous==2.0.1
Jinja2==3.0.1
